Debug
Still not fixed the issue of "undefined reference to `uncompress'"

+++ src/BGEN.cpp
@@ -14,18 +14,16 @@
 #include <memory>
 #include <time.h>
 #include <stdint.h>
-#include <zlib.h>
+// #include <zlib.h>
 
-#include <boost/iostreams/filter/zstd.hpp>
+// #include <boost/iostreams/filter/zstd.hpp>
+#include "zstd.h"
+#include "zlib.h"
 #include <boost/date_time.hpp>
 
 #include <Rcpp.h>
 
-
-
-
 #include "BGEN.hpp"
-
 
 namespace BGEN {
 
@@ -63,13 +61,13 @@ namespace BGEN {
 
     /********** READ HEADER v1.2**********/
     m_fin = fopen(t_bgenFileName.c_str(), "rb");
-    uint offset; fread(&offset, 4, 1, m_fin); //cout << "offset: " << offset << endl;
-    uint L_H; fread(&L_H, 4, 1, m_fin); //cout << "L_H: " << L_H << endl;
-    uint m_M0; fread(&m_M0, 4, 1, m_fin); std::cout << "snpBlocks (Mbgen): " << m_M0 << std::endl;
+    unsigned int offset; fread(&offset, 4, 1, m_fin); //cout << "offset: " << offset << endl;
+    unsigned int L_H; fread(&L_H, 4, 1, m_fin); //cout << "L_H: " << L_H << endl;
+    unsigned int m_M0; fread(&m_M0, 4, 1, m_fin); std::cout << "snpBlocks (Mbgen): " << m_M0 << std::endl;
     assert(Mbgen != 0);
-    //uint Nbgen; fread(&Nbgen, 4, 1, m_fin); std::cout << "samples (Nbgen): " << Nbgen << std::endl;
+    //unsigned int Nbgen; fread(&Nbgen, 4, 1, m_fin); std::cout << "samples (Nbgen): " << Nbgen << std::endl;
     fread(&m_N0, 4, 1, m_fin); std::cout << "samples (Nbgen): " << m_N0 << std::endl;
-    uint m_Nsample = t_SampleInBgen.size();
+    unsigned int m_Nsample = t_SampleInBgen.size();
     m_SampleInBgen = t_SampleInBgen;
     if (m_N0 != m_Nsample) {
       std::cerr << "ERROR: Number of samples in BGEN header does not match sample file" << std::endl;
@@ -77,10 +75,10 @@ namespace BGEN {
     }
     char magic[5]; fread(magic, 1, 4, m_fin); magic[4] = '\0'; //cout << "magic bytes: " << string(magic) << endl;
     fseek(m_fin, L_H-20, SEEK_CUR); //cout << "skipping L_H-20 = " << L_H-20 << " bytes (free data area)" << endl;
-    uint flags; fread(&flags, 4, 1, m_fin); //cout << "flags: " << flags << endl;
-    uint CompressedSNPBlocks = flags&3; std::cout << "CompressedSNPBlocks: " << CompressedSNPBlocks << std::endl;
+    unsigned int flags; fread(&flags, 4, 1, m_fin); //cout << "flags: " << flags << endl;
+    unsigned int CompressedSNPBlocks = flags&3; std::cout << "CompressedSNPBlocks: " << CompressedSNPBlocks << std::endl;
     assert(CompressedSNPBlocks==1); // REQUIRE CompressedSNPBlocks==1
-    uint Layout = (flags>>2)&0xf; std::cout << "Layout: " << Layout << std::endl;
+    unsigned int Layout = (flags>>2)&0xf; std::cout << "Layout: " << Layout << std::endl;
     assert(Layout==1 || Layout==2); // REQUIRE Layout==1 or Layout==2
     fseek(m_fin, offset+4, SEEK_SET);
   }
@@ -113,7 +111,7 @@ namespace BGEN {
   }
 
 
-void BgenClass::Parse2(unsigned char *buf, uint bufLen, const unsigned char *zBuf, uint zBufLen,std::string & snpName,std::vector< double > & dosages, double & AC, double & AF, std::vector<int> & indexforMissing, double & info, std::vector<unsigned int> & iIndex) {
+void BgenClass::Parse2(unsigned char *buf, unsigned int bufLen, const unsigned char *zBuf, unsigned int zBufLen,std::string & snpName,std::vector< double > & dosages, double & AC, double & AF, std::vector<int> & indexforMissing, double & info, std::vector<unsigned int> & iIndex) {
 
   uLong destLen = bufLen;
   if (uncompress(buf, &destLen, zBuf, zBufLen) != Z_OK || destLen != bufLen) {
@@ -122,43 +120,43 @@ void BgenClass::Parse2(unsigned char *buf, uint bufLen, const unsigned char *zBu
   }
 
   unsigned char *bufAt = buf;
-    uint N = bufAt[0]|(bufAt[1]<<8)|(bufAt[2]<<16)|(bufAt[3]<<24); bufAt += 4;
+    unsigned int N = bufAt[0]|(bufAt[1]<<8)|(bufAt[2]<<16)|(bufAt[3]<<24); bufAt += 4;
 
   if (N != m_N0) {
       std::cerr << "ERROR: " << snpName << " has N = " << N << " (mismatch with header block)" << std::endl;
       exit(1);
     }
-    uint K = bufAt[0]|(bufAt[1]<<8); bufAt += 2;
+    unsigned int K = bufAt[0]|(bufAt[1]<<8); bufAt += 2;
     if (K != 2U) {
       std::cerr << "ERROR: " << snpName << " has K = " << K << " (non-bi-allelic)" << std::endl;
       exit(1);
     }
-    uint Pmin = *bufAt; bufAt++;
+    unsigned int Pmin = *bufAt; bufAt++;
     if (Pmin != 2U) {
       std::cerr << "ERROR: " << snpName << " has minimum ploidy = " << Pmin << " (not 2)" << std::endl;
       exit(1);
     }
-    uint Pmax = *bufAt; bufAt++;
+    unsigned int Pmax = *bufAt; bufAt++;
     if (Pmax != 2U) {
       std::cerr << "ERROR: " << snpName << " has maximum ploidy = " << Pmax << " (not 2)" << std::endl;
       exit(1);
     }
 
     const unsigned char *ploidyMissBytes = bufAt;
-    for (uint i = 0; i < N; i++) {
-      uint ploidyMiss = *bufAt; bufAt++;
+    for (unsigned int i = 0; i < N; i++) {
+      unsigned int ploidyMiss = *bufAt; bufAt++;
       if (ploidyMiss != 2U && ploidyMiss != 130U) {
         std::cerr << "ERROR: " << snpName << " has ploidy/missingness byte = " << ploidyMiss
              << " (not 2 or 130)" << std::endl;
         exit(1);
       }
     }
-    uint Phased = *bufAt; bufAt++;
+    unsigned int Phased = *bufAt; bufAt++;
     if (Phased != 0U) {
       std::cerr << "ERROR: " << snpName << " has Phased = " << Pmax << " (not 0)" << std::endl;
       exit(1);
     }
-    uint B = *bufAt; bufAt++;
+    unsigned int B = *bufAt; bufAt++;
     if (B != 8U) {
       std::cerr << "ERROR: " << snpName << " has B = " << B << " (not 8)" << std::endl;
       exit(1);
@@ -176,7 +174,7 @@ void BgenClass::Parse2(unsigned char *buf, uint bufLen, const unsigned char *zBu
     }
     std::size_t missing_cnt = 0;
 
-        for (uint i = 0; i < N; i++) {
+        for (unsigned int i = 0; i < N; i++) {
      //if(i == 1){std::cout << "ploidyMissBytes[i] " << ploidyMissBytes[i] << std::endl;}
      if (ploidyMissBytes[i] != 130U){
       //bufAt += 2;
@@ -254,40 +252,40 @@ void BgenClass::Parse2(unsigned char *buf, uint bufLen, const unsigned char *zBu
   {
     if(t_fileStartPos > 0){fseek(m_fin, t_fileStartPos, SEEK_SET);}
      std::string SNPID, RSID, chromosome, first_allele,second_allele ;
-     uint position;
+     unsigned int position;
      std::vector< std::string > alleles ;
      std::vector< double > dosages;
      double AC, AF, info;
      std::vector< int > indexforMissing;
-     std::vector< uint > iIndex;
+     std::vector< unsigned int > iIndex;
      char snpID[65536], rsID[65536], chrStr[65536];
-     uint maxLA = 65536, maxLB = 65536;
+     unsigned int maxLA = 65536, maxLB = 65536;
      char *allele1, *allele0;
      allele1 = (char *) malloc(maxLA+1);
      allele0 = (char *) malloc(maxLB+1);
-     ushort LS; size_t numBoolRead = fread(&LS, 2, 1, m_fin); // cout << "LS: " << LS << " " << std::flush;
+     uint32_t LS; size_t numBoolRead = fread(&LS, 2, 1, m_fin); // cout << "LS: " << LS << " " << std::flush;
      bool isBoolRead;
      Rcpp::List result ;
      if ( numBoolRead > 0 ) {
       isBoolRead = true;
       fread(snpID, 1, LS, m_fin); snpID[LS] = '\0'; // cout << "snpID: " << string(snpID) << " " << std::flush;
-      ushort LR; fread(&LR, 2, 1, m_fin); // cout << "LR: " << LR << " " << std::flush;
+      uint32_t LR; fread(&LR, 2, 1, m_fin); // cout << "LR: " << LR << " " << std::flush;
       fread(rsID, 1, LR, m_fin); rsID[LR] = '\0'; // cout << "rsID: " << string(rsID) << " " << std::flush;
       RSID = std::string(rsID)=="." ? snpID : rsID;
       //std::string SNPID = string(snpID);
 
-      ushort LC; fread(&LC, 2, 1, m_fin); // cout << "LC: " << LC << " " << std::flush;
+      uint32_t LC; fread(&LC, 2, 1, m_fin); // cout << "LC: " << LC << " " << std::flush;
       fread(chrStr, 1, LC, m_fin); chrStr[LC] = '\0';
       chromosome  = std::string(chrStr);
 
-      uint physpos; fread(&physpos, 4, 1, m_fin); // cout << "physpos: " << physpos << " " << std::flush;
+      unsigned int physpos; fread(&physpos, 4, 1, m_fin); // cout << "physpos: " << physpos << " " << std::flush;
       position = physpos;
-      ushort K; fread(&K, 2, 1, m_fin); //cout << "K: " << K << endl;
+      uint32_t K; fread(&K, 2, 1, m_fin); //cout << "K: " << K << endl;
       if (K != 2) {
         std::cerr << "ERROR: Non-bi-allelic variant found: " << K << " alleles" << std::endl;
         exit(1);
       }
-      uint LA; fread(&LA, 4, 1, m_fin); // cout << "LA: " << LA << " " << std::flush;
+      unsigned int LA; fread(&LA, 4, 1, m_fin); // cout << "LA: " << LA << " " << std::flush;
       if (LA > maxLA) {
         maxLA = 2*LA;
         free(allele1);
@@ -295,7 +293,7 @@ void BgenClass::Parse2(unsigned char *buf, uint bufLen, const unsigned char *zBu
       }
       fread(allele1, 1, LA, m_fin); allele1[LA] = '\0';
       second_allele = std::string(allele1);
-      uint LB; fread(&LB, 4, 1, m_fin); // cout << "LB: " << LB << " " << std::flush;
+      unsigned int LB; fread(&LB, 4, 1, m_fin); // cout << "LB: " << LB << " " << std::flush;
       if (LB > maxLB) {
         maxLB = 2*LB;
         free(allele0);
@@ -304,10 +302,10 @@ void BgenClass::Parse2(unsigned char *buf, uint bufLen, const unsigned char *zBu
       fread(allele0, 1, LB, m_fin); allele0[LB] = '\0';
       first_allele = std::string(allele0);
 
-      uint C; fread(&C, 4, 1, m_fin); //cout << "C: " << C << endl;
+      unsigned int C; fread(&C, 4, 1, m_fin); //cout << "C: " << C << endl;
       if (C > m_zBuf.size()) m_zBuf.resize(C-4);
       //std::cout << "m_zBuf.size() " << m_zBuf.size() << std::endl;
-      uint D; fread(&D, 4, 1, m_fin); //cout << "D: " << D << endl;
+      unsigned int D; fread(&D, 4, 1, m_fin); //cout << "D: " << D << endl;
       m_zBufLens = C-4; m_bufLens = D;
       fread(&m_zBuf[0], 1, C-4, m_fin);
       AC = 0;
